ezpass/requirement.txt
selenium
pyodbc
pandas
bs4
requests
pytesseract
Pillow